+++ Sources/MapboxSearchUI/SearchCategoriesRootView.swift
@@ -102,6 +102,42 @@ class SearchCategoriesRootView: UIView {
         categoriesTableView.separatorColor = configuration.style.separatorColor
     }
 
+    override func layoutSubviews() {
+        super.layoutSubviews()
+
+        /// Make sure that RTL users display the default tab
+        /// The scroll view will start at content offset (0, 0)
+        /// but the start tab will register as favorites.
+        /// When starting out with RTL, the current tab is favorites and != to `.categories` tab,
+        /// then manually set the content offset to the appropriate `.categories` default tab.
+        if effectiveUserInterfaceLayoutDirection == .rightToLeft {
+            let page = contentScrollView.contentOffset.x / contentScrollView.bounds.width
+            let currentTab = CategoriesFavoritesSegmentControl.Tab(
+                scrollViewPageProgress: page,
+                direction: effectiveUserInterfaceLayoutDirection
+            )
+
+            let defaultTab = CategoriesFavoritesSegmentControl.Tab.categories
+            guard currentTab != defaultTab else {
+                return
+            }
+
+            contentScrollView.contentOffset.x = defaultTab.horizontalOffsetFor(scrollView: contentScrollView)
+
+            // Without forcing a refresh the titles and masks will not display correctly (invisible or grayed-out)
+            // Force another layout pass to ensure these display correctly.
+            segmentedControl.setNeedsLayout()
+            segmentedControl.setNeedsDisplay()
+            segmentedControl.layoutIfNeeded()
+
+            // On first-draw we have just assigned the tab to the default and we know
+            // that this will render incorrectly for RTL users.
+            // Re-assigning the progress to the (backwards) location of the second tab
+            // (really "first tab" (which is zero-indexed)) will force it to redraw correctly.
+            segmentedControl.selectionSegmentProgress = 1
+        }
+    }
+
     func resetUI(animated: Bool) {
         contentScrollView.setContentOffset(.zero, animated: animated)
     }
@@ -139,7 +175,10 @@ extension SearchCategoriesRootView: FavoritesTableViewSourceDelegate {
 extension SearchCategoriesRootView: UIScrollViewDelegate {
     func scrollViewDidScroll(_ scrollView: UIScrollView) {
         let page = scrollView.contentOffset.x / scrollView.bounds.width
-        let newTab = CategoriesFavoritesSegmentControl.Tab(scrollViewPageProgress: page)
+        let newTab = CategoriesFavoritesSegmentControl.Tab(
+            scrollViewPageProgress: page,
+            direction: scrollView.effectiveUserInterfaceLayoutDirection
+        )
 
         if segmentedControl.selectedTab != newTab {
             segmentedControl.selectedTab = newTab
@@ -159,7 +198,8 @@ extension SearchCategoriesRootView {
             .beginFromCurrentState,
             .allowUserInteraction,
             .curveEaseInOut,
-        ], animations: {
+        ], animations: { [weak self] in
+            guard let self else { return }
             self.contentScrollView.contentOffset.x = self.segmentedControl.selectedTab.horizontalOffsetFor(
                 scrollView:
                 self.contentScrollView
@@ -169,20 +209,37 @@ extension SearchCategoriesRootView {
 }
 
 extension CategoriesFavoritesSegmentControl.Tab {
-    fileprivate init(scrollViewPageProgress: CGFloat) {
-        if scrollViewPageProgress <= 0.5 {
-            self = .categories
+    /// When the scroll view is instantiated and at the default location, show the categories Tab
+    ///
+    fileprivate init(scrollViewPageProgress: CGFloat, direction: UIUserInterfaceLayoutDirection) {
+        if direction == .leftToRight {
+            if scrollViewPageProgress <= 0.5 {
+                self = .categories
+            } else {
+                self = .favorites
+            }
         } else {
-            self = .favorites
+            /// Right to Left behavior
+            if scrollViewPageProgress <= 0.5 {
+                self = .favorites
+            } else {
+                self = .categories
+            }
         }
     }
 
     fileprivate func horizontalOffsetFor(scrollView: UIScrollView) -> CGFloat {
-        switch self {
-        case .categories:
+        switch (self, scrollView.effectiveUserInterfaceLayoutDirection) {
+        case (.categories, .leftToRight):
             return 0
-        case .favorites:
+        case (.favorites, .leftToRight):
             return scrollView.bounds.width
+        case (.categories, .rightToLeft):
+            return scrollView.bounds.width
+        case (.favorites, .rightToLeft):
+            return 0
+        case (_, _):
+            fatalError("Unsupported text direction")
         }
     }
 }